+++ src/test/clojure/clara/test_clear_ns_vars.clj
@@ -1,7 +1,7 @@
-;;; Tests that clear-ns-productions! correction clears all vars marked as productions from the namespace.
-(ns clara.test-clear-ns-productions
+;;; Tests that clear-ns-vars! correction clears all vars marked as productions from the namespace.
+(ns clara.test-clear-ns-vars
   (:require
-   [clara.rules :refer [clear-ns-productions! defquery defrule defsession
+   [clara.rules :refer [clear-ns-vars! defquery defrule defsession
                         fire-rules insert insert! query]]
    [clara.tools.testing-utils :as tu]
    [clojure.test :refer [deftest is testing use-fixtures]])
@@ -20,14 +20,14 @@
 
 (def ^:production-seq ns-production-seq-to-be-cleared
   [{:doc  "Before clearing"
-    :name "clara.test-clear-ns-productions/production-seq-to-be-cleared"
+    :name "clara.test-clear-ns-vars/production-seq-to-be-cleared"
     :lhs  '[{:type        :a
              :constraints []}]
     :rhs  '(clara.rules/insert! :before-clearing-seq)}])
 
-(defsession uncleared-session 'clara.test-clear-ns-productions :fact-type-fn identity)
+(defsession uncleared-session 'clara.test-clear-ns-vars :fact-type-fn identity)
 
-(clear-ns-productions!)
+(clear-ns-vars!)
 
 (defrule rule-after-clearing
   [:a]
@@ -41,12 +41,12 @@
 
 (def ^:production-seq production-seq-after-clearing
   [{:doc  "After clearing"
-    :name "clara.test-clear-ns-productions/production-seq-after-clearing"
+    :name "clara.test-clear-ns-vars/production-seq-after-clearing"
     :lhs  '[{:type        :a
              :constraints []}]
     :rhs  '(clara.rules/insert! :after-clearing-seq)}])
 
-(defsession cleared-session 'clara.test-clear-ns-productions :fact-type-fn identity)
+(defsession cleared-session 'clara.test-clear-ns-vars :fact-type-fn identity)
 
 ;;; Then tests validating what productions the respective sessions have.
 (deftest cleared?
@@ -54,7 +54,7 @@
     (is (= :before-clearing @tu/side-effect-holder))
     (reset! tu/side-effect-holder nil))
   (let [cleared (-> cleared-session (insert :a) (fire-rules))]
-    (testing "cleared-session should not contain any productions before (clear-ns-productions!)"
+    (testing "cleared-session should not contain any productions before (clear-ns-vars!)"
       (is (= nil @tu/side-effect-holder))
       (is (empty? (query cleared query-before-clearing)))
       (is (not-empty (query cleared query-after-clearing))))
@@ -64,6 +64,6 @@
 (deftest query-cleared?
   (let [uncleared (-> uncleared-session (insert :a) (fire-rules))
         cleared (-> cleared-session (insert :a) (fire-rules))]
-    (is (not-empty (query uncleared "clara.test-clear-ns-productions/query-to-be-cleared")))
-    (is (thrown-with-msg? IllegalArgumentException #"clara.test-clear-ns-productions/query-to-be-cleared"
-                          (query cleared "clara.test-clear-ns-productions/query-to-be-cleared")))))
+    (is (not-empty (query uncleared "clara.test-clear-ns-vars/query-to-be-cleared")))
+    (is (thrown-with-msg? IllegalArgumentException #"clara.test-clear-ns-vars/query-to-be-cleared"
+                          (query cleared "clara.test-clear-ns-vars/query-to-be-cleared")))))